DPL: refactor metrics helpers
* Move helpers in separate file.
* Move some template invariant code to separate out-of-line
  function.
* Separate booking of the metric from cursor creation.
