+++ Framework/Core/CMakeLists.txt
@@ -52,6 +52,7 @@ o2_add_library(Framework
                        src/DataSpecUtils.cxx
                        src/DeviceConfigInfo.cxx
                        src/DeviceMetricsInfo.cxx
+                       src/DeviceMetricsHelper.cxx
                        src/DeviceSpec.cxx
                        src/DeviceController.cxx
                        src/DeviceSpecHelpers.cxx
